Move `initialize_tensor` function to utils module and update imports
