+++ tests/test_functional.py
@@ -139,13 +139,5 @@ class TestMeshgrid2D(unittest.TestCase):
         self.assertTrue(torch.allclose(actual_y, expected_y))
 
 
-class TestInitializetensor(unittest.TestCase):
-    def test_initialize_tensor(self):
-        with self.assertRaises(ValueError):
-            initialize_tensor("name", 1.0, is_scalar=True, is_complex=True, is_integer=True)
-        with self.assertRaises(ValueError):
-            initialize_tensor("name", 1.1, is_scalar=True, is_complex=False, is_integer=True)
-
-
 if __name__ == "__main__":
     unittest.main()
